@@ -10,29 +10,29 @@
     }
   ],
   "require": {
-    "php": ">=8.1",
-    "doctrine/annotations": "^1.14",
+    "php": ">=8.2",
+    "doctrine/annotations": "^2.0",
     "doctrine/doctrine-bundle": "^2.8",
-    "doctrine/orm": "^2.14",
+    "doctrine/orm": "^2.2",
     "elasticsearch/elasticsearch": "^7.1",
-    "symfony/dotenv": "6.4.*",
-    "symfony/expression-language": "6.4.*",
-    "symfony/framework-bundle": "^6.4",
-    "symfony/messenger": "6.4.*",
-    "symfony/property-info": "6.4.*",
-    "symfony/runtime": "6.4.*",
-    "symfony/serializer": "6.4.*",
-    "symfony/validator": "6.4.*",
-    "symfony/yaml": "6.4.*",
+    "symfony/dotenv": "7.2.*",
+    "symfony/expression-language": "7.2.*",
+    "symfony/framework-bundle": "^7.2",
+    "symfony/messenger": "7.2.*",
+    "symfony/property-info": "7.2.*",
+    "symfony/runtime": "7.2.*",
+    "symfony/serializer": "7.2.*",
+    "symfony/validator": "7.2.*",
+    "symfony/yaml": "7.2.*",
     "ext-simplexml": "*",
-    "symfony/security-bundle": "^6.4",
-    "symfony/finder": "^6.4"
+    "symfony/security-bundle": "^7.2",
+    "symfony/finder": "^7.2"
   },
   "require-dev": {
-    "symfony/browser-kit": "6.4.*",
-    "symfony/css-selector": "6.4.*",
-    "symfony/console": "6.4.*",
-    "symfony/phpunit-bridge": "^6.4",
+    "symfony/browser-kit": "7.2.*",
+    "symfony/css-selector": "7.2.*",
+    "symfony/console": "7.2.*",
+    "symfony/phpunit-bridge": "^7.2",
     "symfony/test-pack": "^1.1"
   },
   "prefer-stable": true,
@@ -48,7 +48,7 @@
       "dev-master": "1.0-dev"
     },
     "symfony": {
-      "require": "6.4.*"
+      "require": "7.2.*"
     }
   },
   "conflict": {
